pyproject.toml
[project]
name = "eli5-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "openai>=2.33.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
